numeric labels start from 0 and saved as floats instead of int

+++ src/load_data/folder_loader.hpp
@@ -8,7 +8,7 @@ struct ImageData {
     //cv::Mat image;
     std::string filePath;
     std::string classStr;
-    int label;
+    float label;
 };
 
 std::vector<ImageData> loadImgsFromFolder(const std::string& folderPath);
